Moved NamedAccounts class to Cluster.
